Time: 58 ms (41.94%), Space: 27.4 MB (92.52%) - LeetHub

+++ 704-binary-search/704-binary-search.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int search(vector<int>& arr, int target)
     {
-        int n = (int) arr.size();
+        int n =  arr.size();
         
         int low=0;
         int high= n-1;
